requirements.txt
a2a-sdk[http-server]>=0.3.25
httpx>=0.27.2
uvicorn>=0.30.6
python-multipart>=0.0.12
